actions: Split out zerver.actions.custom_profile_fields.
