+++ zerver/lib/actions.py
@@ -34,6 +34,7 @@ from typing_extensions import TypedDict
 from analytics.lib.counts import COUNT_STATS, do_increment_logging_stat
 from confirmation import settings as confirmation_settings
 from confirmation.models import Confirmation, create_confirmation_link, generate_key
+from zerver.actions.custom_profile_fields import do_remove_realm_custom_profile_fields
 from zerver.actions.default_streams import (
     do_remove_default_stream,
     do_remove_streams_from_default_stream_group,
@@ -78,7 +79,6 @@ from zerver.lib.exceptions import (
     StreamWithIDDoesNotExistError,
     ZephyrMessageAlreadySentException,
 )
-from zerver.lib.external_accounts import DEFAULT_EXTERNAL_ACCOUNTS
 from zerver.lib.i18n import get_language_name
 from zerver.lib.markdown import MessageRenderingResult, topic_links
 from zerver.lib.markdown import version as markdown_version
@@ -154,7 +154,7 @@ from zerver.lib.topic import (
     update_edit_history,
     update_messages_for_topic_edit,
 )
-from zerver.lib.types import EditHistoryEvent, ProfileDataElementValue, ProfileFieldData
+from zerver.lib.types import EditHistoryEvent
 from zerver.lib.upload import delete_avatar_image
 from zerver.lib.user_counts import realm_user_count, realm_user_count_by_role
 from zerver.lib.user_groups import (
@@ -179,8 +179,6 @@ from zerver.models import (
     ArchivedAttachment,
     Attachment,
     Client,
-    CustomProfileField,
-    CustomProfileFieldValue,
     DefaultStream,
     DefaultStreamGroup,
     Draft,
@@ -208,7 +206,6 @@ from zerver.models import (
     active_non_guest_user_ids,
     active_user_ids,
     bot_owner_user_ids,
-    custom_profile_fields_for_realm,
     get_bot_dicts_in_realm,
     get_bot_services,
     get_client,
@@ -6478,172 +6475,6 @@ def do_remove_realm_domain(
         do_set_realm_property(realm, "emails_restricted_to_domains", False, acting_user=acting_user)
     event = dict(type="realm_domains", op="remove", domain=domain)
     transaction.on_commit(lambda: send_event(realm, event, active_user_ids(realm.id)))
-
-
-def notify_realm_custom_profile_fields(realm: Realm) -> None:
-    fields = custom_profile_fields_for_realm(realm.id)
-    event = dict(type="custom_profile_fields", fields=[f.as_dict() for f in fields])
-    send_event(realm, event, active_user_ids(realm.id))
-
-
-def try_add_realm_default_custom_profile_field(
-    realm: Realm, field_subtype: str
-) -> CustomProfileField:
-    field_data = DEFAULT_EXTERNAL_ACCOUNTS[field_subtype]
-    custom_profile_field = CustomProfileField(
-        realm=realm,
-        name=field_data["name"],
-        field_type=CustomProfileField.EXTERNAL_ACCOUNT,
-        hint=field_data["hint"],
-        field_data=orjson.dumps(dict(subtype=field_subtype)).decode(),
-    )
-    custom_profile_field.save()
-    custom_profile_field.order = custom_profile_field.id
-    custom_profile_field.save(update_fields=["order"])
-    notify_realm_custom_profile_fields(realm)
-    return custom_profile_field
-
-
-def try_add_realm_custom_profile_field(
-    realm: Realm,
-    name: str,
-    field_type: int,
-    hint: str = "",
-    field_data: Optional[ProfileFieldData] = None,
-) -> CustomProfileField:
-    custom_profile_field = CustomProfileField(realm=realm, name=name, field_type=field_type)
-    custom_profile_field.hint = hint
-    if (
-        custom_profile_field.field_type == CustomProfileField.SELECT
-        or custom_profile_field.field_type == CustomProfileField.EXTERNAL_ACCOUNT
-    ):
-        custom_profile_field.field_data = orjson.dumps(field_data or {}).decode()
-
-    custom_profile_field.save()
-    custom_profile_field.order = custom_profile_field.id
-    custom_profile_field.save(update_fields=["order"])
-    notify_realm_custom_profile_fields(realm)
-    return custom_profile_field
-
-
-def do_remove_realm_custom_profile_field(realm: Realm, field: CustomProfileField) -> None:
-    """
-    Deleting a field will also delete the user profile data
-    associated with it in CustomProfileFieldValue model.
-    """
-    field.delete()
-    notify_realm_custom_profile_fields(realm)
-
-
-def do_remove_realm_custom_profile_fields(realm: Realm) -> None:
-    CustomProfileField.objects.filter(realm=realm).delete()
-
-
-def try_update_realm_custom_profile_field(
-    realm: Realm,
-    field: CustomProfileField,
-    name: str,
-    hint: str = "",
-    field_data: Optional[ProfileFieldData] = None,
-) -> None:
-    field.name = name
-    field.hint = hint
-    if (
-        field.field_type == CustomProfileField.SELECT
-        or field.field_type == CustomProfileField.EXTERNAL_ACCOUNT
-    ):
-        field.field_data = orjson.dumps(field_data or {}).decode()
-    field.save()
-    notify_realm_custom_profile_fields(realm)
-
-
-def try_reorder_realm_custom_profile_fields(realm: Realm, order: List[int]) -> None:
-    order_mapping = {_[1]: _[0] for _ in enumerate(order)}
-    custom_profile_fields = CustomProfileField.objects.filter(realm=realm)
-    for custom_profile_field in custom_profile_fields:
-        if custom_profile_field.id not in order_mapping:
-            raise JsonableError(_("Invalid order mapping."))
-    for custom_profile_field in custom_profile_fields:
-        custom_profile_field.order = order_mapping[custom_profile_field.id]
-        custom_profile_field.save(update_fields=["order"])
-    notify_realm_custom_profile_fields(realm)
-
-
-def notify_user_update_custom_profile_data(
-    user_profile: UserProfile, field: Dict[str, Union[int, str, List[int], None]]
-) -> None:
-    data = dict(id=field["id"], value=field["value"])
-
-    if field["rendered_value"]:
-        data["rendered_value"] = field["rendered_value"]
-    payload = dict(user_id=user_profile.id, custom_profile_field=data)
-    event = dict(type="realm_user", op="update", person=payload)
-    send_event(user_profile.realm, event, active_user_ids(user_profile.realm.id))
-
-
-def do_update_user_custom_profile_data_if_changed(
-    user_profile: UserProfile,
-    data: List[Dict[str, Union[int, ProfileDataElementValue]]],
-) -> None:
-    with transaction.atomic():
-        for custom_profile_field in data:
-            field_value, created = CustomProfileFieldValue.objects.get_or_create(
-                user_profile=user_profile, field_id=custom_profile_field["id"]
-            )
-
-            # field_value.value is a TextField() so we need to have field["value"]
-            # in string form to correctly make comparisons and assignments.
-            if isinstance(custom_profile_field["value"], str):
-                custom_profile_field_value_string = custom_profile_field["value"]
-            else:
-                custom_profile_field_value_string = orjson.dumps(
-                    custom_profile_field["value"]
-                ).decode()
-
-            if not created and field_value.value == custom_profile_field_value_string:
-                # If the field value isn't actually being changed to a different one,
-                # we have nothing to do here for this field.
-                continue
-
-            field_value.value = custom_profile_field_value_string
-            if field_value.field.is_renderable():
-                field_value.rendered_value = render_stream_description(
-                    custom_profile_field_value_string
-                )
-                field_value.save(update_fields=["value", "rendered_value"])
-            else:
-                field_value.save(update_fields=["value"])
-            notify_user_update_custom_profile_data(
-                user_profile,
-                {
-                    "id": field_value.field_id,
-                    "value": field_value.value,
-                    "rendered_value": field_value.rendered_value,
-                    "type": field_value.field.field_type,
-                },
-            )
-
-
-def check_remove_custom_profile_field_value(user_profile: UserProfile, field_id: int) -> None:
-    try:
-        custom_profile_field = CustomProfileField.objects.get(realm=user_profile.realm, id=field_id)
-        field_value = CustomProfileFieldValue.objects.get(
-            field=custom_profile_field, user_profile=user_profile
-        )
-        field_value.delete()
-        notify_user_update_custom_profile_data(
-            user_profile,
-            {
-                "id": field_id,
-                "value": None,
-                "rendered_value": None,
-                "type": custom_profile_field.field_type,
-            },
-        )
-    except CustomProfileField.DoesNotExist:
-        raise JsonableError(_("Field id {id} not found.").format(id=field_id))
-    except CustomProfileFieldValue.DoesNotExist:
-        pass
 
 
 def do_update_outgoing_webhook_service(
